[#2] refactor: 불필요한 중복 코드 파일 삭제
