+++ src/features/main/league-adaptation/ui/AdaptationList.tsx
@@ -1,4 +1,4 @@
-import AdaptationItem from '../../../player-search/ui/AdaptionItem'
+import AdaptationItem from '../../../player-search/ui/AdaptationItem'
 import type { AdaptationListProps } from '../model/type'
 
 export default function AdaptationList({ currentData }: AdaptationListProps) {
@@ -24,7 +24,7 @@ export default function AdaptationList({ currentData }: AdaptationListProps) {
           <div className="flex flex-col gap-1.5">
             {currentData.players.map((player, i) => (
               <AdaptationItem
-                key={player.name}
+                key={player.id}
                 player={player}
                 rank={i + 1}
                 animationDelay={i * 60}
