+++ src/pages/main/ui/MainPage.tsx
@@ -3,8 +3,8 @@ import HeroBanner from './HeroBanner'
 import { SearchBar } from '@/features/player-search'
 import { useState } from 'react'
 import StatsSection from './StatsSection'
-import LeagueAdaptationSection from '@/features/main/league-adaptation/ui/LeagueAdaptionSection'
-import FeatureHighlightSection from './FeatureHightlightSection'
+import LeagueAdaptationSection from '@/features/main/league-adaptation/ui/LeagueAdaptationSection'
+import FeatureHighlightSection from './FeatureHighlightSection'
 import CtaBanner from '@/features/main/league-adaptation/ui/CtaBanner'
 
 export function MainPage({ isHeaderVisible = true }: MainPageProps) {
