--- src/shared/lib/topLeague.ts
@@ -1,1 +1,0 @@
-export type TopLeagueKey = 'EPL' | 'LA' | 'BL' | 'SA' | 'L1'
